Radiotap: deal with extended present flags
My atheros dongle dongle actually pushes frames with extended present
flags. This requires the parser to, at least, skip the extra present
flags in order to continue parsing from the right place.

Incidentally this arose another problem: the radiotap header specifications
requires each field to be aligned to the start of the packet, and when
there are two present flags bitmap, this is even less obvious. In particular
the TSFT ended up to be improperly aligned here.

This patch makes the radiotap parser to take care of both there issues, and
fixes wrong shift/mask for extended present flag bit.