Use `six` package for backwards compatibility
The `six` package provides a number of backwards compatible hooks for
moved/renamed APIs. It's better to use their APIs than reinvent the wheel.

While this doesn't completely remove the need for dpkt/compat.py, it
largely eliminates this need. Removing/replacing `compat_ord` and
remaining backwards compatible between python 2.7 and python 3.7 proved
to be more of a challenge than I was willing to commit to for this
change.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>