Implement PktlenReader class
This class extends pcap Reader by exposing len field from packet header
to the user. This provides information about the original packet size as
it was transmitted. For trimmed packets with corrupted len fields there is
actually no other way to know the read packet size.

The main purpose of PktlenReader however is to be used in conjunction with
Writer.writepkt()'s pktlen argument to preserve this pcap packet header
field when working on files generated with tcpdump's --snapshot-length.

Authored-by: Filipp Mikoian <filipp@u.nus.edu>